Split keymaster identity routes
