+++ services/keymaster/server/src/keymaster-api.ts
@@ -14,11 +14,16 @@ import WalletCache from '@didcid/keymaster/wallet/cache';
 import CipherNode from '@didcid/cipher/node';
 import { InvalidParameterError } from '@didcid/common/errors';
 import config from './config.js';
+import { createAddressRouter } from './keymaster-address-router.js';
+import { createAddressTailRouter } from './keymaster-address-tail-router.js';
 import { createRequireAdminKey } from './keymaster-admin.js';
 import { createContentRouter } from './keymaster-content-router.js';
 import { createCoreRouter } from './keymaster-core-router.js';
 import { createCredentialRouter } from './keymaster-credential-router.js';
+import { createDidCommRouter } from './keymaster-didcomm-router.js';
 import { createIdentityRouter } from './keymaster-identity-router.js';
+import { createLightningRouter } from './keymaster-lightning-router.js';
+import { createNostrRouter } from './keymaster-nostr-router.js';
 import { createPublicRouter } from './keymaster-public-router.js';
 import type { CreateKeymasterRouterOptions } from './keymaster-router-types.js';
 import promClient from 'prom-client';
@@ -160,6 +165,11 @@ v1router.use(createPublicRouter(routerOptions));
 v1router.use(createRequireAdminKey(config));
 v1router.use(createCoreRouter(routerOptions));
 v1router.use(createIdentityRouter(routerOptions));
+v1router.use(createAddressRouter(routerOptions));
+v1router.use(createDidCommRouter(routerOptions));
+v1router.use(createAddressTailRouter(routerOptions));
+v1router.use(createNostrRouter(routerOptions));
+v1router.use(createLightningRouter(routerOptions));
 v1router.use(createCredentialRouter(routerOptions));
 v1router.use(createContentRouter(routerOptions));
