+++ swaggerConf.js
@@ -45,6 +45,11 @@ const keymasterOptions = {
         'services/keymaster/server/src/keymaster-public-router.ts',
         'services/keymaster/server/src/keymaster-core-router.ts',
         'services/keymaster/server/src/keymaster-identity-router.ts',
+        'services/keymaster/server/src/keymaster-address-router.ts',
+        'services/keymaster/server/src/keymaster-didcomm-router.ts',
+        'services/keymaster/server/src/keymaster-address-tail-router.ts',
+        'services/keymaster/server/src/keymaster-nostr-router.ts',
+        'services/keymaster/server/src/keymaster-lightning-router.ts',
         'services/keymaster/server/src/keymaster-credential-router.ts',
         'services/keymaster/server/src/keymaster-content-router.ts'
     ]
